requirements.txt
Django~=3.2.25
django-extensions~=3.2.3
git+https://github.com/dbca-wa/ledger#egg=ledger
git+https://github.com/dbca-wa/dpaw-utils.git@0.4.2#egg=dpaw-utils
#git+https://github.com/django-oscar/django-oscar.git@8a3288da439cc2a878f44ae5c5101043e658d2a2#egg=django-oscar
social-auth-app-django~=2.1.0
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage~=7.8.0
coveralls~=4.0.1
reportlab~=4.4.1
django_bootstrap3~=23.6
django-datatables-view==1.20.0
django-reversion~=4.0.2
django-reversion-compare~=0.17.0
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
django-countries~=7.6.1
django-cron~=0.6.0
django-dynamic-fixture~=4.0.1
openpyxl~=3.1.5
#datapackage==0.8.1
py4j~=0.10.9.9
djangorestframework~=3.16.0
djangorestframework-gis~=1.1
djangorestframework-csv~=3.0.2
pycountry~=24.6.1
six>=1.10.0
django-ical~=1.9.2
django-taggit~=4.0.0
geojson~=3.2.0
#unicodecsv==0.14.1
PyPDF2~=3.0.1
#django-extra-fields==0.9
django-crispy-forms~=1.14.0
django-reset-migrations~=0.4.0
django-ckeditor~=6.7.2
djangorestframework-datatables~=0.7.2
django-confy==1.0.4
#xlwt==1.3.0
#xlsxwriter==1.1.2
gunicorn~=23.0.0
#dj-static==0.0.6
mixer~=7.2.2
django-dirtyfields~=1.9.7
#dj-database-url==0.5.0
python-docx~=1.1.2
docxtpl~=0.20.0
phonenumbers~=9.0.5
django-treebeard~=4.7.1
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
django-multiselectfield~=0.1.13
psycopg2~=2.9.10
deepdiff~=8.5.0
python-dateutil~=2.8.2
numpy~=2.2.6
pandas~=2.2.3
whitenoise~=5.3.0
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv